lab_1/requirements.txt
sqlmodel
fastapi
psycopg2-binary
alembic
python-dotenv
passlib
python-jose
bcrypt==3.1.7